simulator robot now turns around obstacle
Because of the wall sensor not being the same in the simulator as on the actual robot, the simulator robot just bumps into an obstacle, thinks its stuck in a corner, and rotate counterclockwise about 10 degrees and then turns a corner on repeat.  This does seem to move the robot around the obstacle, but need to test on actual robot if it works as intended with correct wall sensors.